refactor for principal app and display a heroes list
link files for app and create a mock-heroes for list
