+++ src/app/principalApp/app.module.ts
@@ -3,8 +3,8 @@ import { FormsModule } from '@angular/forms';
 import { BrowserModule } from '@angular/platform-browser';
 
 import { AppRoutingModule } from './app-routing.module';
-import { AppComponent } from './principalApp/app.component';
-import { HeroesComponent } from './components/heroes/heroes.component';
+import { AppComponent } from '../principalApp/app.component'; //module app principal
+import { HeroesComponent } from '../components/heroes/heroes.component'; //module heroes created
 
 @NgModule({
   declarations: [AppComponent, HeroesComponent],
